폼 숫자 체크확인하는 로직 추가

+++ app/src/main/kotlin/com/acc/goodwill/data/source/ContributorDao.kt
@@ -42,7 +42,7 @@ import javax.inject.Singleton
         registrationNumber: String?,
         join: Int
     ) {
-        var id: EntityID<Int>? = null
+        var id: EntityID<Long>? = null
         val launchResult = suspendedTransactionAsync(Dispatchers.IO) {
             id = ContributorTable.insertAndGetId {
                 it[this.name] = name
